+++ src/cli/index.ts
@@ -10,6 +10,7 @@ import configCommand from "@/commands/config";
 import mentionsCommand from "@/commands/mentions";
 import activityCommand from "@/commands/activity";
 import notificationsCommand from "@/commands/notifications";
+import prCommand from "@/commands/pr";
 import { GhitgudError } from "@/core/errors";
 
 const NAME = "ghitgud";
@@ -24,6 +25,7 @@ mentionsCommand.register(program);
 pingCommand.register(program);
 labelsCommand.register(program);
 configCommand.register(program);
+prCommand.register(program);
 
 program.addHelpText("before", ascii);
 program.exitOverride();